+++ SBVUnitTest/SBVUnitTestBuildTime.hs
@@ -2,4 +2,4 @@
 module SBVUnitTestBuildTime (buildTime) where
 
 buildTime :: String
-buildTime = "Sat Dec 12 09:49:10 PST 2015"
+buildTime = "Wed Dec 16 19:23:47 PST 2015"