+++ CHANGES.md
@@ -8,6 +8,9 @@
   * Export HasKind(..) from the Dynamic interface. Thanks to Adam Foltzer for the patch.
   * More careful handling of SMT-Lib reserved names.
   * Update tested version of MathSAT to 5.3.9
+  * Generalize sShiftLeft/sShiftRight/sRotateLeft/sRotateRight to work with signed
+    shift/rotate amounts, where negative values revert the direction. Similar
+    generalizations are also done for the dynamic variants.
 
 ### Version 5.6, 2015-12-06